refactor: 불필요한 코드 제거 및 파일 이동
