+++ docs/drop.sql
@@ -6,4 +6,4 @@ DROP TABLE TBL_ADVERTISEMENTS;
 DROP TABLE TBL_OAUTH2_ROLE;
 DROP TABLE TBL_OAUTH2_MEMBER;
 DROP TABLE TBL_MEMBER;
-DROP TABLE TBL_PRODUCT_KEYWORDS
+DROP TABLE TBL_PRODUCT_KEYWORDS;
