+++ src/main/resources/mybatis/mybatis-mapper/cartMapper.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="com.everypet.cart.mapper.CartMapper">
+<mapper namespace="com.everypet.cart.model.dao.CartMapper">
     <insert id="insertCart" parameterType="Cart">
         INSERT INTO TBL_CART (
             MEMBER_ID, PRODUCT_ID, CART_QUANTITY
