+++ src/main/resources/mybatis/mybatis-mapper/memberMapper.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="com.everypet.member.mapper.MemberMapper">
+<mapper namespace="com.everypet.member.model.dao.MemberMapper">
 
     <select id="selectMemberByMemberId" parameterType="String" resultType="Member">
         SELECT *
