+++ src/main/resources/mybatis/mybatis-mapper/roleMapper.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.everypet.member.mapper.RoleMapper">
+<mapper namespace="com.everypet.member.model.dao.RoleMapper">
 
     <insert id="insertRole" parameterType="String">
         INSERT INTO TBL_ROLE(MEMBER_ID, AUTHORITIES) values (#{memberId}, #{authorities})
