+++ src/test/java/com/everypet/member/model/mapper/MemberDaoTest.java
@@ -1,7 +1,7 @@
 package com.everypet.member.model.mapper;
 
 import com.everypet.global.config.DatabaseConfig;
-import com.everypet.member.mapper.MemberMapper;
+import com.everypet.member.model.dao.MemberMapper;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
